Extract DefaultsBridges into its own file
